Modified the mover code artist/performer
some files from Amazon use the artist ID3 field to contain a list of all the
performers on the track, and use the 'performer' field for the name of the
album artist.

Modified so that when both artist and performer are present, we use the
shorter of the two fields as the artist name.