History file
When we move/copy a directory of files, we also now add a file named

YYYY_Album_Name.history

that contains the timestamp that the directory was created there. If we move
things multiple times, we append those timestamps to the end. So if that file
contains:

100
233
401

...it was added to our system at t=100, and the most recent move/copy was
performed at t=401