added new --musiconly option
When true, we only move/copy music files, we don't handle anything else that
happens to be in those directories (jpg files, etc.). Useful when copying
to a location (USB drive for the car) that can't possibly use those files, and
recovers at least a little space on the drive.