Checking for last modification time in place.
Unless we're forcing a full scan, we only process track files that are in
directories that have been modified since the last time we looked at them
(which in practice will almost always be new directories since the last time
we did a scan). Forcing a scan will always look at everything.